api_rest_user_managment_modul/requirements.txt
asgiref==3.7.2
Django==4.2.9
django-recaptcha==4.0.0
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
